# perf(sql): optimize memory usage in hash joins