--- a/core/src/main/java/io/questdb/griffin/engine/join/HashJoinLightRecordCursorFactory.java
+++ b/core/src/main/java/io/questdb/griffin/engine/join/HashJoinLightRecordCursorFactory.java
@@ -24,7 +24,10 @@
 
 package io.questdb.griffin.engine.join;
 
-import io.questdb.cairo.*;
+import io.questdb.cairo.CairoConfiguration;
+import io.questdb.cairo.ColumnTypes;
+import io.questdb.cairo.RecordSink;
+import io.questdb.cairo.TableToken;
 import io.questdb.cairo.map.Map;
 import io.questdb.cairo.map.MapFactory;
 import io.questdb.cairo.map.MapKey;
@@ -49,7 +52,7 @@ public class HashJoinLightRecordCursorFactory extends AbstractJoinRecordCursorFa
             RecordCursorFactory masterFactory,
             RecordCursorFactory slaveFactory,
             @Transient ColumnTypes joinColumnTypes,
-            @Transient ColumnTypes valueTypes, // this expected to be just 3 INTs, we store chain references in map
+            @Transient ColumnTypes valueTypes, // this expected to be just 2 INTs, we store chain references in map
             RecordSink masterSink,
             RecordSink slaveKeySink,
             int columnSplit,
@@ -119,12 +122,36 @@ public class HashJoinLightRecordCursorFactory extends AbstractJoinRecordCursorFa
                 key.put(masterRecord, keySink);
                 MapValue value = key.findValue();
                 if (value != null) {
-                    size += value.getInt(2);
+                    size += value.getInt(1);
                 }
             }
             return size;
         } finally {
             masterCursor.toTop();
+        }
+    }
+
+    private static void populateRowIDHashMap(
+            SqlExecutionCircuitBreaker circuitBreaker,
+            RecordCursor cursor,
+            Map keyMap,
+            RecordSink recordSink,
+            LongChain rowIDChain
+    ) {
+        final Record record = cursor.getRecord();
+        while (cursor.hasNext()) {
+            circuitBreaker.statefulThrowExceptionIfTripped();
+
+            MapKey key = keyMap.withKey();
+            key.put(record, recordSink);
+            MapValue value = key.createValue();
+            if (value.isNew()) {
+                value.putInt(0, rowIDChain.put(record.getRowId(), -1));
+                value.putInt(1, 1);
+            } else {
+                value.putInt(0, rowIDChain.put(record.getRowId(), value.getInt(0)));
+                value.addInt(1, 1);
+            }
         }
     }
 
@@ -224,7 +251,7 @@ public class HashJoinLightRecordCursorFactory extends AbstractJoinRecordCursorFa
 
         private void buildMapOfSlaveRecords() {
             if (!isMapBuilt) {
-                TableUtils.populateRowIDHashMap(circuitBreaker, slaveCursor, joinKeyMap, slaveKeySink, slaveChain);
+                populateRowIDHashMap(circuitBreaker, slaveCursor, joinKeyMap, slaveKeySink, slaveChain);
                 isMapBuilt = true;
             }
         }