… core/src/main/java/io/questdb/griffin/engine/join/LtJoinLightRecordCursorFactory.java
… core/src/main/resources/io/questdb/site/conf/server.conf
… core/src/test/java/io/questdb/test/PropServerConfigurationTest.java
… core/src/test/java/io/questdb/test/ServerMainTest.java
… core/src/test/java/io/questdb/test/griffin/GroupByTest.java
… core/src/test/java/io/questdb/test/griffin/JoinTest.java
… core/src/test/java/io/questdb/test/griffin/SqlOptimiserTest.java
… core/src/test/java/io/questdb/test/griffin/engine/join/LongChainTest.java
… core/src/test/java/io/questdb/test/griffin/engine/window/WindowFunctionTest.java
… pkg/ami/marketplace/assets/server.conf