Optimizes memory usage in hash join factories that use random access (`*Light*` factories):
* `LongChain` now manages its heap and stores compressed offsets (`int`s) similarly to `OrderedMap`. Each value stored in the chain now takes 12 bytes instead of 16 bytes.
* Inner hash join map's values are now 2 `int`s instead of 3 `long`s. Outer hash join map's values are single `int`.

**Behavior change**: `LongChain` now stores reversed list of long values (row ids). This impacts inner/outer join result set order in case of chained right-hand rows, but since joins in SQL have no order guarantees and we already use unordered maps to handle hash joins, it shouldn't be a breaking change for our users.

These changes give a small speed-up in hash join exec times:

```sql
-- with id=x%2600000:
--   master: 27s
--   patch: 22.5s
-- with id=x:
--   master: 54s
--   patch: 46s
select count(*)
from (
  select *
  from t1
  inner join t2 on t1.id = t2.id
);

create table t1 as (
select
  x%2600000 as id,
  timestamp_sequence(0, 100000) ts 
from long_sequence(140000000)
) timestamp(ts) PARTITION BY year;

create table t2 as (
select
  x%2600000 as id,
  timestamp_sequence(0, 100000) ts 
from long_sequence(140000000)
) timestamp(ts) PARTITION BY year;
```