@ideoma 2024-08-21
#### [PR Coverage check]

:heart_eyes: **pass** : 87 / 88 (98.86%)




#### file detail

|   |path|covered line|new line|coverage|
|----|----|----|----|----|
|:large_blue_circle:|io/questdb/griffin/engine/join/LongChain.java|37|38|97.37%|
|:large_blue_circle:|io/questdb/griffin/engine/join/HashOuterJoinFilteredRecordCursorFactory.java|1|1|100.00%|
|:large_blue_circle:|io/questdb/cairo/TableUtils.java|6|6|100.00%|
|:large_blue_circle:|io/questdb/griffin/engine/join/HashOuterJoinLightRecordCursorFactory.java|1|1|100.00%|
|:large_blue_circle:|io/questdb/griffin/engine/join/HashJoinRecordCursorFactory.java|14|14|100.00%|
|:large_blue_circle:|io/questdb/griffin/SqlCodeGenerator.java|8|8|100.00%|
|:large_blue_circle:|io/questdb/griffin/engine/join/HashOuterJoinFilteredLightRecordCursorFactory.java|5|5|100.00%|
|:large_blue_circle:|io/questdb/griffin/engine/join/HashOuterJoinRecordCursorFactory.java|1|1|100.00%|
|:large_blue_circle:|io/questdb/griffin/engine/join/HashJoinLightRecordCursorFactory.java|14|14|100.00%|