@ideoma 2024-08-22
#### [PR Coverage check]

:heart_eyes: **pass** : 105 / 106 (99.06%)




#### file detail

|   |path|covered line|new line|coverage|
|----|----|----|----|----|
|:large_blue_circle:|io/questdb/griffin/engine/join/LongChain.java|37|38|97.37%|
|:large_blue_circle:|io/questdb/griffin/engine/join/HashOuterJoinFilteredRecordCursorFactory.java|1|1|100.00%|
|:large_blue_circle:|io/questdb/griffin/engine/join/HashOuterJoinLightRecordCursorFactory.java|13|13|100.00%|
|:large_blue_circle:|io/questdb/griffin/engine/join/HashJoinRecordCursorFactory.java|14|14|100.00%|
|:large_blue_circle:|io/questdb/griffin/SqlCodeGenerator.java|7|7|100.00%|
|:large_blue_circle:|io/questdb/griffin/engine/join/HashOuterJoinFilteredLightRecordCursorFactory.java|4|4|100.00%|
|:large_blue_circle:|io/questdb/griffin/engine/join/HashOuterJoinRecordCursorFactory.java|1|1|100.00%|
|:large_blue_circle:|io/questdb/griffin/engine/join/HashJoinLightRecordCursorFactory.java|28|28|100.00%|